+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+* Support for a "Superuser" defined in config that can use all commands
+
 ### Fixed
 * Long lists of names are now split to avoid exceeding character limits
 * Direct messages sent to the bot don't crash it any more